VideoCapture: abolish OnFrameInfo almost everywhere.
Historically OnFrameInfo was a one-time message that gave the
resolution of all subsequent frames. This patch eliminates that, to
allow the resolution to be specified on a frame-by-frame basis, at the
device's discretion.

At the IPC layer two changes were necessary, with the rest of the code
falling in line:  [1] Resolution/layout info (in the form of a
VideoCaptureFormat) is sent with each frame delivery.  [2] Shared-
memory buffers need to be reallocated, so it was necessary to add a
message to drop a buffer.

A significant rework of VideoCaptureBufferPool handles a shift to a
world where buffers are allocated/re-allocated on demand, and where
the buffers in the pool may not all be the same size at any given
moment. This allows the buffer pool to be allocated earlier, at the
start of the lifetime of the controller.

OnFrameInfo methods are gone everywhere but the very beginning of the
pipeline (between the Device and its Client) and the very end (an
OnFrameInfo event is synthesized in the renderer process for the
benefit of the pepper video capture host).

BUG=266082

Review URL: https://codereview.chromium.org/23551011