Add Views-specific download danger prompt dialog
Create a Views web contents modal dialog implementation of the download
danger prompt dialog, so that we can apply bold text formatting to the
lead text in the dialog. Adds a TODO for the same change on other platforms.

See the associated bug for a screenshot.

BUG=305387

Review URL: https://codereview.chromium.org/28673003