[mac] Support adding the launcher to the Dock when Dashboard.app is present.
Dashboard has a special tile-type of "dashboard-tile" (rather than
"file-tile"), and no file-data, when it is present in a user's Dock.
dock::AddIcon(..) currently emits an error when encountering these dock
tiles.

This CL will detect the case of a Dock tile with no file-data, and
substitute a dummy path (the empty string) to align with the source data
dictionary.

For the App Launcher, this CL also adds some error handling for
dock::AddIcon. This allows the app launcher creation code to pop up the
Finder window when adding the icon to the Dock fails.

BUG=309941
TEST=Drag "Dashboard" from /Applications to the Dock, and enable the app
launcher in a fresh Chrome install. You should get an app launcher added
to the Dock.

Review URL: https://codereview.chromium.org/32563004