Split AwSettings contruction out of AwContents
BUG=305352
Internal b/10932261

This allows the AwSettings (java only) instance to be created in advance
of AwContents, thereby allowing the AwSettings to be used early on when
the chromium stack has not be fully started.
As a result AwSettings is now able to live as a 'dumb' Java container of
settings, prior to the chromium stack being initialized at all.

Change-Id: I4059a93a7ad132043998fe2eab318a3b066c882e

Review URL: https://codereview.chromium.org/26959003