ozone: Allow the output to be written to a file.
If ozone is started with --ozone-dump-file flag, then the output of the display
is dumped to a file in PNG format. The path of the file can be specified as
well, for example: "--ozone-dump-file=/tmp/ozone.png". /dev/null can be used as
the path to avoid any real file left over in the system, but still run through
this custom surface-factory implementation.

BUG=none
R=dnicoara@chromium.org, rjkroege@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/32193012