cc: Fix touch hit-testing for overlapping layers (in tests).
If a touch-layer is below a non-touch layer, and the two layers overlap,
then the overlapped region should not be treated as a touch-handling region.
This change updates FindLayerThatIsHitByPointInTouchHandlerRegion() to do the
right thing (which is currently used only in tests), and uses this function from
LayerTreeHostImpl::HaveTouchEventHandlersAt() (instead of having its own code).

BUG=none
R=enne@chromium.org

Review URL: https://codereview.chromium.org/27718006