sync: Force LOCAL GUs if invalidations disabled
If the invalidations system is not working or has been not working for
some time in the recent past, we should not allow the syncer to skip
making the "LOCAL" request for updates prior to each commit request.

Currently, all pre-commit update avoidance is hidden behind a flag, so
this change will have no effect on normal users.

BUG=309815

Review URL: https://codereview.chromium.org/33723003