Files.app: Change when to send BEGIN progress event.
Previously tasks of FileOperationManager is queued and the BEGIN progress event
is sent when the task reaches to the top of the queue.  But currently we can
show the multiple tasks by using ProgressCenter.  This CL let
FileOperationManager send the BEGIN event when the task is added to the queue.

BUG=277432
TEST=manually

Review URL: https://codereview.chromium.org/36603002