Simple http fallback for the data reduction proxy.
This CL adds a new command line option and preprocessor macro for a second proxy that the data reduction proxy feature will add to its configs as a fallback from the primary proxy. If the fallback proxy isn't defined, only the primary proxy is part of the configurations, and vice-versa. 

(Also fixes a bug where a http proxy specified on the command line would still be configured as HTTPS in the generated PAC file). 

BUG=306634

Review URL: https://codereview.chromium.org/30883003