[rAC, OSX] Fix keyView handling.
The rAC dialog used to try to restore whatever the previous keyview was
once the loading shield ended - that triggered inadvertently autopopup
messages.

This change removes the firstResponder hack that did this to hide focus
rings and instead does the right thing - hide the underlying dialog
when the shield is showing.

BUG=310724
R=rsesek@chromium.org

Review URL: https://codereview.chromium.org/39273003