Browser test for PINless authentication.
This test verifies that a PIN is required before pairing and is not required
afterwards. It then cleans up the pairing and tries again. The second run
serves to verify that the clean-up worked correctly and also allows us to
test both methods of clearing a pairing.

Review URL: https://codereview.chromium.org/38903002