[rAC, OSX] Properly update key view loop.
Recalculate the key view loop for rAC dialog whenever layout changes.
A global layout change is indicative of a size/visibility change of
individual sections of the dialog, which in turn means that there is
potentially a different visible set of key views that need to be re-
connected in a new loop.

BUG=308170
R=isherman@chromium.org

Review URL: https://codereview.chromium.org/40003003