FileAPI: Make ObfuscatedFileUtil to receive SandboxFSDelegate in its ctor
After introducing PluginPrivateFSB, each SandboxFSBDelegate and the backend has
its own ObfuscatedFileUtil, and the util should not depend on the delegate when
it is owned by the backend.

This change makes the ctor of the util to receive a pointer of the delegate and
to check its existence when accessing it.


BUG=286240
TEST=content_unittests --gtest_filter=ObfuscatedFileUtilTest.*

Review URL: https://codereview.chromium.org/42713002