This changelist does three things:
Fixes the scrollbar problem mentioned in the bug. The kExternalInstallLeftColumnWidth is too wide (350px) for the dialog (385px) to show, since it also needs a 13 px padding and 69px of icon space. But, it doesn't make sense to specify an absolute width for the dialog given that the left column can vary greatly in size, so now we calculate the width based on the columns it has, so that people don't exceed the available space (previously specified as a const).

Also added some padding to the right of the image, to make the image align visually with the button in the dialog (looks much better) and made the extension icon image right-aligned within the field.

And finally, fixed the rowspan for the case when the extension has no special permissions (it should span 3 rows in that case, not 4).

BUG=292646
R=sky@chromium.org

Review URL: https://codereview.chromium.org/39113004