Reland: Implement features in NativeViewHostAura for scroll end effect
This CL enables the fast resize path on Aura by implementing
NativeViewHostAura::InstallClip. It also adds the concept of gravity to
the fast resize state to control the positioning of the clip relative
to the contents. Gravity is implemented for Aura only in this CL and
ignored in other implementations.

Refactored |clipping_window_| so that ownership is not in NVHA.
Added test to make sure that the leak is resolved.
Fixed occlusion issue with find bar by setting kHostViewKey for
|clipping_window_|.

BUG=151356
BUG=282463
BUG=308075
TEST=Confirmed the fast path and gravity works in conjunction with
     scroll end effect CL.
     Added and updated unit tests.
     Ran ash_unittests and views_unittests locally under headchecker.
     Ran ash_unittests and views_unittests locally under ASAN.

Review URL: https://codereview.chromium.org/30993004