Add support for forcing isolated testing on builder level.
Add a flag '--force-isolated' for runisolatedtest.py that
enforces the test to run isolated, ignoring which test or
builder name is used.

Add a lookup for the factory property 'force_isolated' in
runtest.py that will pass this flag to runisolatedtest.py.

TEST=Testing on a local slave checkout, running with the
flag and verified isolate.py gets used for the execution.
BUG=

Review URL: https://codereview.chromium.org/41993005