Add 'quick-check' option to install-build-deps.sh
This is a lightweight check to alert developers when their build
dependencies are out-of-date, and hopefully avoid confusing compile
errors. It takes < 1 second to run, so can be checked whenever sources
are synced, and might even by worth including as a default step in
gyp_chromium or DEPS hooks.

By definition, this is not a complete check, and in particular, will not
detect version updates of installed dependencies. It's mostly intended
to detect when entirely new dependencies have been added.

R=markus@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/40603004