net:  Rework how blocking delegates are logged.
Switch to logging an event before calling into delegates, and after
they return instead of only in the case that they block.  This allows
delegates to log events as they take turns blocking a network
request.

Also allows delegates to log events when delaying the start of
a request, and when delaying the start of reading the body of a
request (But not between normal body reads).

Also allows delegates to log delegate information that will not be
displayed to the user.

BUG=294920

Review URL: https://codereview.chromium.org/25097004