cc: RenderPass::CopyAll should not break if quads are culled.
If quads are culled from more than one consecutive layer, then multiple
shared quad states may appear in the list but not be used. CopyAll
should not break in this scenario, so use a loop instead of an if to
to find the corrent shared quad state.

This is validated by the cc_messages IPC de-pickling already, and
it allows this valid IPC through, but then we fail to deal
with it in CopyAll.

Tests:
RenderPassTest.CopyAllWithCulledQuads

R=piman
BUG=311438

Review URL: https://codereview.chromium.org/45033002