Pass weakptr bound callbacks to MediaSourceDelegate
Most of the demuxer callbacks are on the media loop, and MediaSourceDelegate will post the callback on to the main thread.
Passing weakptrs for safety.
Currently only update_network_state_cb needs the weakptr. So this CL is an overkill but just for consistency purpose.

BUG=311316

Review URL: https://codereview.chromium.org/41923005