First step of new component updater on-demand framework: direct API.
This is the first in a series of CLs to implement the new on-demand
framework for the component update service. I've temporarily
'friended' the PNaCl installer so that I can make those changes in a
separate changelist; once we reach consensus on the triggering
mechanism.

This is a re-upload of issue 27000013.

BUG=307193

Review URL: https://codereview.chromium.org/37943003