Land Recent QUIC changes.
Rename ambiquious ConnectionClose method to OnConnectionClosed.

Merge internal change: 54774694

Change TCP cubic's max packet size used to calculate the initial and
subsequent congestion window in bytes to the default of 1460 from QUIC's
artificially low 1200.

Merge internal change: 54772582

QUIC: make the SCID a hash of the rest of the server config.

We need to ensure that the SCID changes whenever the rest of the server
config does. For example, cl/54004815 changed the server config, but not
the SCID. This could lead to bad 0-RTT handshakes where the server
believes the handshake is good, but will derive different keys than the
client.

The easiest change to ensure that the SCID is changed is to make it a
hash of the rest of the message.

Merge internal change: 54659325

EndToEndTest to fix TSAN.

Merge internal change: 54090381

Don't add a zero-length public key to kPUBS if P-256 support is
disabled. kKEXS and kPUBS should have the same number of elements.

Merge internal change: 54004815

Make the QuicConnectionHelper no longer owned by the QuicConnection.
Instead a single helper is owned by the Dispatcher or Client, and shared
among each Connection.

Merge internal change: 53974968

R=rch@chromium.org

Review URL: https://codereview.chromium.org/45733002