Remember value of IsImplSidePaintingEnabled
Command line arguments are constant at runtime, but the function
cc::switches::IsImplSidePaintingEnabled() does two searches through
the list of arguments every time it's called, and it's called every
time we record a SkPicture.

With this patch, the function remembers the value determined by the
first call and avoids repeating the searches on subsequent calls.

BUG=311248
R=nduca@chromium.org,skyostil@chromium.org

Review URL: https://codereview.chromium.org/43463003