Add WaitForInitialUserActivity policy
This CL adds a user policy that controls whether power management delays
and the session length limit should only start running after initial user
activity in a session. The SessionLengthLimiter is modified to obey the
new policy. Power management will hooked up to the new policy in a
follow-up CL.

It would technically be possible to add separate policies that determine
whether power management delays on the one hand and the session length
limit on the other hand should wait for initial user activity. However,
it does not make sense for one of these to start running on session start
while the other waits for user activity. Therefore, a single policy is
added for both.

BUG=310176
TEST=Updated unit and browser tests

Review URL: https://codereview.chromium.org/35943002