Enable apps that use Push Messaging to not request background mode.
The use case: the app is using Push Messaging for not real-time event that happens rarely. There is no great reason to  keep Chrome running just for that.

Implemented as a whitelist for now since we believe this is a temporary solution. A better solution, which would allow push to work all the time but not keep Chrome running can take longer time. See bug for more info.

BUG=311268

Review URL: https://codereview.chromium.org/45863002