cc: Fix DCHECK in PictureLayer::PushPropertiesTo
The DCHECK will only hold if PictureLayer::Update is called
in that frame. DrawContents is not enough since there are
other cases in LayerTreeHostCommon::LayerShouldBeSkipped
that causes Update to be skipped.

Instead keep track of the frame number when Update was
and compare it to the current frame number.

BUG=311311

Review URL: https://codereview.chromium.org/42653002