Fix PrerenderBrowserTestWithExtensions.WebNavigation test.
It got broken after it was disabled. Fix it and bring its disabled configs
in line with other extension browser_test flakes. This is on the theory
that this is the only source of the flake. If the test still fails, it can
be disabled again.

Also remove a chrome.tabs.update call in
PrerenderBrowserTestWithExtensions.TabsApi that wasn't doing anything.

BUG=177163,249179
TEST=PrerenderBrowserTestWithExtensions.*

Review URL: https://codereview.chromium.org/28273007