Force recreating the session service for session restore after shutdown.
Normally, when the session service is shut down, it cannot be recreated
for a profile. This prevents the session service from being recreated to
restore a session and resume session tracking if a browser is re-opened for a
profile without first quitting Chrome.

This change recreates the session service after shutdown if it is to be
used for session restore.

BUG=268224

Review URL: https://codereview.chromium.org/26110003