Extract ipclist target from chrome.gyp.
The gyp files under chrome/ are really not well organized, kind of a mess,
so it is hard to find the right places to put things. As most of the things goes
into browser, browser_ui or common libraries, this isn't much an issue though.
But we can do better.

BUG=None
TEST=build/gyp_chromium chrome/tools/ipclist/ipclist.gyp
     ninja -C out/Debug ipclist
R=thestig@chromium.org

Review URL: https://codereview.chromium.org/46853003