Pepper: Add in_process() to BrowserPpapiHostImpl.
This is required as part of the FileIO change. BrowserPpapiHostImpl now serves
as a host for in-process plugins in the renderer. Calling
GetPluginProcessHandle() is unsafe if that field isn't set, so this change adds
an in_process() method so that callers can know whether they can safely call
GetPluginProcessHandle().

BUG=246396
R=bbudge@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/49223007