Convert extensions::LazyBackgroundTaskQueue from Profile to BrowserContext
To do this, introduce the concept of an ExtensionsBrowserClient that allows
the extensions module to ask Profile-like questions of the embedder. The
rest of the change is mostly mechanical.

Also, introduce some conventions around profile() vs. browser_context() and
GetForProfile() vs. GetForBrowserContext() so we can tell which pieces of
the system have been converted.

BUG=309909
TEST=browser_tests LazyBackgroundPageApiTest.* unit_tests LazyBackgroundTaskQueueTest.*

Review URL: https://codereview.chromium.org/46793003