Mac: delay window raising until space transitions.
When manually raising browser windows in response to a dock activation that
triggered a space transition, isOnActiveSpace is wrong and we focus the wrong
space's windows. Detect this by seeing if the key window is off-screen. If so,
delay the raise until the space change notification.

BUG=309656
TEST=see bug

Review URL: https://codereview.chromium.org/37253004