Added flag to enable AVFoundation in Mac
From the flag documentation:
" AVFoundation is available in versions 10.7 and onwards, and is to be used
http://crbug.com/288562 for both audio and video device monitoring and for
video capture. Being a dynamically loaded NSBundle and library, it hits the
Chrome startup time (http://crbug.com/311325 and http://crbug.com/311437);
until development is finished and the library load time issue is solved, the
usage of this library is hidden behind this flag. "

BUG=288562, 311325, 311437

Review URL: https://codereview.chromium.org/50213004