Disable parallel execution for some WebRTC browser tests
The changes in http://crrev.com/231359 made all tests being
executed in parallel. This works for most tests, but not for
the WebRTC browser tests that are relying on the peerconnection
server for setting up a WebRTC video call between two tabs.

Adding  --test-launcher-jobs=1 to the test arguments should make
all the tests for this step to execut in sequence according to
phajdan.jr@.

TEST=none
BUG=none

Review URL: https://codereview.chromium.org/47923012