Plumb a testing-only API for forcing resizes.
In the new world of unicorns and rainbow, the layout tests
will actually use the same machinery as the rest of Chromium
for window resizing, but there is still one ... uhm,
forcing function that still needs to be invoked at the beginning
of each test to eliminate any possibilities of tests affecting
each other. This function is RenderWidget::ResizeSynchronously,
with complementary plumbing.

BUG=309760
R=jam, jochen, tony

Review URL: https://codereview.chromium.org/47973002