linux_aura: Don't implement InputStateLookup.
In X11, we only receive events of mouse events; we cannot query the
state ourselves. The best we can do here is use the built in event
tracking to emulate it. This is not perfect.

This enhances the tab dragging code (which runs mostly outside of
aura's event handling stuff) so that it sets the global Env mouse
button data when it starts and stops, so this should fix the case
of tab dragging.

BUG=306048

Review URL: https://codereview.chromium.org/50713006