Fix cropping around the header.
Before, when the image was behind the gallery's header, it was imposible to drag around this area. This patch disables the header when it gets dimmed, so it doesn't drag but forwards events to the crop tool. Therefore, since now the behaviour for the header and the footer is the same.

Along the way, the draggable header has been updated to reflect recently changed window buttons.

TEST=Tested manually. Make the image fill the entire window, and try to crop under the footer and under the header.
BUG=307823

Review URL: https://codereview.chromium.org/50483002