Disable thread under DRT content_shell unless explicitly enabled
We happen to not enable threaded compositing when running
layout tests by virtue of the fact that it is enabled by a Finch
experiment which is only initialized in chrome.

When getting rid of this Finch experiment, a bunch of layout
tests blew up because that had the effect of enabling threaded
compositing on the tests. This makes it so that won't happen.

BUG=302738, 233830, 306557

Review URL: https://codereview.chromium.org/51333003