[Telemetry] Fix page sets with both file:// and http:// URLs on Android.
When we start the browser, it is configured to always use WPR (this makes http
work). When we go to hit the file:// URL, we need to hit the other web server
instead of WPR, but we can't reconfigure Chrome without restarting.

So, WPR has logic to forward any 127.0.0.1 requests on to another server to
allow the local HTTP server to get the requests. The problem was that we were
explicitly requesting localhost instead of 127.0.0.1 which fooled WPR.

BUG=311582

Review URL: https://codereview.chromium.org/51803002