On iOS, make sure to normalize email address if not already.
Prior to M26, when the user signed in to chrome, it was possible that a
non-normalized email address was set as the autheticated username in
SigninManager.  Now that chrome on iOS normalizes the email, SigninManager
should make sure to save the normalized email address.

BUG=303775

Review URL: https://codereview.chromium.org/50263008