Add |position| to AppListItemModel and add AppListItemList (Take 2)
This adds a |position_| member to AppListItemModel in preparation for
syncing the app list independent of extensions.

It also moves some common code in AppListModel and AppListFolderItem
into AppListItemList, which ensures that |position_| is keept up to date.

This implementation keeps all position maintenance in AppListItemList
and uses AppListItemListObserver to allow ExtensonAppItem to
(temporarily) override the positioning logic.

BUG=305024
For as/shell OWNER:

R=jennyz@chromium.org, xiyuan@chromium.org
TBR=jamescook@chromium.org

Review URL: https://codereview.chromium.org/50003002