Fixes possible use after free in ConflictingModuleViewWin
The problem happens if during ToolbarView::Init() we attempt to show a
bubble in ConflictingModuleViewWin. ConflictingModuleViewWin uses
ToolbarView::app_menu_->GetWidget() as the parent of the bubble. If
app_menu_ is not added at the time we attempt to show the bubble, then
GetWidget() returns NULL and the bubble is not parented to the
browser. This leads to a situation where the browser can be closed
before the bubble, so that when ConflictingModuleViewWin attempts to
use the browser it has since been destroyed.

BUG=309784
TEST=see bug
R=pkasting@chromium.org

Review URL: https://codereview.chromium.org/46673005