Fixing problems with Show / Hide and transient windows for MultiProfile
The problem which is fixed by this issue is that aura::Window::Hide() does hide itself and all transient window children (since https://codereview.chromium.org/10035052), but it does not show all of them. To fix this, the MultiUserWindowManager tracks the visibility for each of the transient children and restores them afterwards.

BUG=307716
TEST=unittests (&visual test)

Review URL: https://codereview.chromium.org/52893002