cc: Don't use LayerTreeHost in LayerTreeTest on impl thread.
Using the LTH is flaky because these function can be called from either
thread, and the LTH may be dropped on the main thread as it is being
called.

In EndTest, we can use ended_ to prevent the function from posting
RealEndTest twice, instead of using the existence of the LayerTreeHost.

Everywhere, use main_task_runner_ instead of
proxy()->MainThreadTaskRunner() which requires accessing the
LayerTreeHost.

R=ajuma
BUG=

Review URL: https://codereview.chromium.org/52993003