WebRequest: Document listeners always getting statusLine
Listeners registered to a couple of the WebRequest API events get a 'details' argument, which always contains the 'statusLine' argument.

It used to be marked as optional, but it's actually always set. This CL drops the optional flag, and adds more explanation about what is the value. This way extension developers may rely on details.statusLine being defined.

BUG=284017

Review URL: https://codereview.chromium.org/52483002