The devsite requires to set up a tree of YAML files that describe the table of contents for the documentation. This was done manually until now, but with our RST infrastructure, it needn't be so. The top-level index.rst, together with the documents themselves have all the info we need to auto-generate these TOC files.
This CL kicks out existing (out-of-date already) manually written TOC files and implements the auto-generation logic instead. The only TOC remnant is for the top-level _book.yaml, which has some "templat-y" info in it. So it's being replaced with _book_template.yaml, from which the final _book.yaml will be generated in the output directory.

Note that the final layout of the devsite is not 100% decided yet, so we may tweak this auto-generation code in the future. I expect this will be easier than tweaking a bunch of hand-maintained YAML files.

BUG=None
R=sbc@chromium.org

Review URL: https://codereview.chromium.org/45333005