[NaCL SDK] Roll naclports revision.
Roll naclports 850 -> 954.  This fixes the accidental
include of the supplemental newlib headers in the ports
bundle.

Most of the changes in the range we are pulling in are
unrelated to the ports that we ship.  A lot of them are
build script related and adding of new ports.  The revisions
of interest are:

 949 - Add port of lua 5.2 and use it in lua_ppapi example.
 948 - Fix accidental modification of top level sdklibs target.

The former changes the default lua version from 5.1 to 5.2
which means we will now be shipping liblua 5.2 in the SDK.
The latter removes the accidentally included headers which
actually fixes bug 313403.

R=binji@chromium.org, binji
BUG=313403

Review URL: https://codereview.chromium.org/52713007