Add better documentation for declarativeContent.
I reorganized the declarative-event common HTML so it's more clear that 3
methods are available.

I moved the CSS condition docs into their own section and elaborated on both the
meaning of a compound selector and the display:none restriction.  It's
unfortunate that this isn't just in the PageStateMatcher.css description, but
that has to live on a single line in the json which makes it really hard to
edit.

BUG=172011
NOTRY=true

Review URL: https://codereview.chromium.org/46533002