CQ: assert instead of re-running verifiers
This is not a valid state for CQ to be in,
and it makes CQ flood Rietveld with comments.

It does not happen under normal operation,
but has happened at least once. Exception
stack trace should help identify how CQ
went there and prevent it from misbehaving.

BUG=none

Review URL: https://codereview.chromium.org/41583004