cc: Fix PicturePile clearing for non-origin canvases
When canvases are larger than content, and content is stretched and
terminates on a non-integer pixel boundary, then the area underneath
that content must be cleared first.  The clearing code attempted to
clear a 2 pixel wide strip, but was not translating it, so it only
worked for canvases located at the origin.

R=brianderson@chromium.org
BUG=296672

Review URL: https://codereview.chromium.org/51433003