Creates functions to make it easier to determine why we're crashing
Currently if creating an HWND fails we crash (similar to out of
memory). There may be a number of reasons why creation fails. This adds
a couple of functions that are called into with common error codes we
see to easily identify the reason.

BUG=82193
TEST=none
R=eroman@chromium.org

Review URL: https://codereview.chromium.org/54913002