Always use chromium_code variable with target_conditions.
The chromium_code variable is designed to be set by targets.
However, if it is used in a conditions predicate in common.gypi,
it will always be false, since conditions are evaluated early.
target_conditions are evaluated late, after the target has had a
chance to set its own variables.

R=mark@chromium.org

Review URL: https://codereview.chromium.org/55283002