Component updater on-demand logic with Resource Throttle
This CL introduces the new way to ask for an on demand install when you have a network request that needs the component to be present:

ComponentUpdaterService::GetOnDemandResourceThrottle()

Returns a resource throttle and initiates a update check + install immediately. The network request associated with the request is unblocked once the component is installed.

The unittests are in progress here
https://codereview.chromium.org/55303002/

BUG=307193
TEST=none
R=darin@chromium.org, sorin@chromium.org, waffles@chromium.org

Review URL: https://codereview.chromium.org/25713007