Make TapCancel ignore ack disposition
There is no reason to block on ack receipt for a TapCancel, particularly as
this event will always reach the main thread and will often precede gestures
that can be handled on the impl thread.  Add TapCancel to the event types
for which event acks will be ignored.

BUG=302852, 313353

Review URL: https://codereview.chromium.org/50433003