[Telemetry] Fix loading_trace measurement with cross-renderer navigations.
The marker dictates which thread's trace events are reported. In the case of a
cross-renderer (e.g. cross-origin) navigation, we want the events of the
thread we ended up on, not the one we started on.

In the future, it would be best to mark and report both.

BUG=
NOTRY=True

Review URL: https://codereview.chromium.org/56153006