Fix an issue that mediaplayer can return size(0,0) for a valid video
After MediaPlayer get prepared, we query the size from it.
However, there are certain cases that the returned size will be (0, 0).
Therefore, we cannot rely on the size info if it is empty.
For fullscreen video, if the returned size is empty, this change will resize the surfaceView to occupy the entire screen space using the onMeasure() call.

BUG=313774

Review URL: https://codereview.chromium.org/56443002