Add a heuristic to try to avoid distorted aspect ratios on Mac.
The test case here is a Logitech B910 opened at 320x240.  QTKit
satisfies this request by opening at 16:9 and downscaling to 4:3, resulting
in an image that is squished (and has severe artifacts due to the low-quality
scaler used in QTKit).

This CL detects that squishing, and retries at 320x180, resulting in an image
with square pixels and no weird scaling artifacts.

BUG=303298

Review URL: https://codereview.chromium.org/25629003