Fixing drag and drop visibility issues of tabs on a visiting desktop
A new browser will be created when a user drags a tab off a tab strip. This browser will automatically get the context of its parent.
When running in multi profile / multiple desktop mode, a browser can visit another user's desktop. In that case that tab will get the ownership of the original browser and made invisible since it belongs to an invisible user. This patch is transferring the visibility of the newly created browser to the current user. Note that this cannot be automatically done since usually browsers get created for "a user" and are automatically visible and owned by "that user" - not the current one. This case is special since it requires that the browser is created from the current user for another user.

While fixing this I have found one problem in the MultiUserWindowManager where the initial show state was not stored and taken into account on transferring to another desktop.

BUG=309793
TEST=unittest & visual test

Review URL: https://codereview.chromium.org/55303003