Mojo should use a command buffer instead of raw GL
This CL switches the native_viewport from using raw GL calls to
using an in-process command buffer. I've plumbed the client
interface to the command buffer as far as the
NativeViewportController, but we really ought to plumb it all
the way into the sample_app.

Of course, once the sample_app moves out-of-process, we'll want
to switch to an out-of-process command buffer.

R=ben@chromium.org, jamesr@chromium.org

Review URL: https://codereview.chromium.org/56543002