Android: Fix calculation of event time for synthetic touch gestures.
On JB+ we use a TimeListener listening to vsync to send out touch events. The time of the touch event was computed by adding the time elapsed since the listener was created to the time when we first touched down a pointer, |mDownTime|. Instead of using |mDownTime| (which is uninitialized the first time round), we record the time when the gesture was started and use that to calculate the event time.

BUG=312731

Review URL: https://codereview.chromium.org/51103011