Record UMA events for actions on the Android NTP.
This records UMA events for actions on the Android NTP (e.g. clicking on
a bookmark), which will be used as part of the time to long dwell
analysis on the NTP. Note: UMA histograms don't contain timestamps, so
we need events instead.

BUG=178932
NOTRY=true
R=mpearson@chromium.org, tedchoc@chromium.org

Review URL: https://codereview.chromium.org/57573002