CQ: do not second-guess the tryserver
Once a try job is triggered through Rietveld, do not submit more jobs
when they look "lost". This overloads tryserver when it is restarted.

I tested this change locally with a dry run (over 24 hours and counting).
I'll coordinate landing this with the trooper and will send a heads-up
e-mail to chrome-troopers.

BUG=none
R=maruel@chromium.org

Review URL: https://codereview.chromium.org/54673016