chromium-status: tests: support split admin console
The newer appengine devserver (1.8.x) launches the admin console on a sep
port rather than a subdir of the main server (like 1.6.x).  This means we
need to find a second free port, and send admin requests to that instead.

Once that's done, we need to update the response parsing slightly:
 - the login page now wants "login" rather than "Login"
 - the interactive console page needs a new field, and its response
   no longer requires parsing of the result (it's raw already)
 - update where & how we get the xsrf_token

BUG=chromium:314501
TEST=`./tests/main_test.py` some tests start actually passing now w/appengine 1.8.6

Review URL: https://codereview.chromium.org/46393005