Docserver: Implement the content providers infrastructure, where a configuration file (content_providers.json) specifies arbitrary places to fetch content from.
As a first pass only the Chromium repository is supported, and our existing 3
content providers (public, exampes, and static) have been moved over to use it.
In a follow-up GitHub will be supported as a content provider along with a
number of other features to content_providers.json so that we're able to host
the rest of Chrome's documentation, in theory.

BUG=293678
R=jyasskin@chromium.org

Review URL: https://codereview.chromium.org/54603010