Implement GetAllDisplays() for the Mac screen backend
Reusing the details of the GetNumDisplays() implementation for returning
a vector of available displays. Reusing the existing code for actually
generating gfx::Display objects by converting to NSScreen and then
calling GetDisplayForScreen to be consistent with the behavior in other
parts of this file.

BUG=311132
R=oshima@chromium.org

Review URL: https://codereview.chromium.org/40163002