Update command state on "load" of a folder
We used to only update the command when selection changes. However, when
a folder is loaded the selection does not change, it goes from [] to []
so no event is fired. However, we do have callbacks after the folder is
populated. Use that to update the state of all the commands.

BUG=313592
R=dbeam@chromium.org

Review URL: https://codereview.chromium.org/57883004