cc: Don't ManageTiles twice in one frame
Any external callers of ManageTiles should inform the Scheduler that
ManageTiles happened so that (1) it can clear the "needs manage tiles"
flag and (2) it can avoid doing a second ManageTiles on the same frame.

Other than after commit, where ManageTiles needs to be called
immediately in order to kick off new raster tasks (or determine that no
raster tasks are required to activate the tree), ManageTiles just needs
to be called periodically to keep the raster jobs working on the most
important content.  Delaying the periodic caller to prevent these bad
frames with commits is a worthwhile tradeoff.

R=brianderson@chromium.org
BUG=314882

Review URL: https://codereview.chromium.org/45923005