extensions: Remove Blacklist dependency from ExtensionPrefs.
ExtensionPrefs, which I'm in the process of moving to the
top-level extensions/ directory, depends on an enum value
from the Blacklist class, which appears to be tied to safe
browsing, a Chrome-specific feature. This change replaces
ExtensionPrefs's use of the enum with a bool.

BUG=313284
TBR=thestig@chromium.org, xiyuan@chromium.org

Review URL: https://codereview.chromium.org/57563006