Deprecate 'singleton' option in chrome.apps.window.create().
This keeps the feature as-is but shows a warning in the developer
console so the developers can fix their code before removing the
feature breaks it.

BUG=313986

TEST=When opening a new window with chrome.apps.window.create() if the
options parameter contain a 'singleton' option set to false and a 'id',
the developer console should show a warning.

Review URL: https://codereview.chromium.org/53973003