Add a SpinningCube to Mojo's sample_app
We don't actually call this code yet, but we're going to need this code in
order to demo a spinning cube. I've verified that this code draws a cube when
called directly from mojo_shell. It should work when called from sample_app
once we figure out how we want to expose GL to Mojo apps.

R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/61423003