third_party/openssl: add ChaCha20+Poly1305 support.
This change is not as scary as it appears. Most of the code has already been
reviewed and is running in production without issues. The only new code is the
ARM support.

ARM now builds both the NEON and generic versions of the code and can enable
the NEON code at runtime by calling CRYPTO_set_NEON_capable(1). This patch does
not contain the code to call that, however.

The addition openssl/patches/channelidchromium.patch and
fix_lhash_iteration.patch is noise from the import script.

BUG=310768

https://codereview.chromium.org/59083010/