Adds a class named Connector that connects a MessagePipe to a MessageReceiver.
MessageReceiver is the interface used by the generated proxies and stubs to transmit messages. Connector is responsible for reading and writing those messages on a MessagePipe. A Connector "is a" MessageReceiver so that it can accept outbound messages, and "has a" MessageReceiver so it can dispatch incoming messages.

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/54743003